# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Edo Talent Hunt API",
    "version": "1.0.0",
    "description": "A comprehensive authentication API for Edo Talent Hunt with email verification, JWT authentication, and password management.",
    "contact": {
      "name": "Edo Talent Hunt Team",
      "email": "tolu.ajuwon@aceall.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://eth-project-backend-1086159474664.europe-west1.run.app",
      "description": "Production server (Google Cloud Run)"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme"
      }
    },
    "schemas": {
...
      "AdminRegistration": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "registrationType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userId": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "personalInfo": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/api/v1/admin/registrations": {
      "get": {
        "summary": "Get all registrations (Admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "submitted",
                "approved",
                "rejected"
              ]
            },
            "description": "Filter by registration status"
          },
          {
            "in": "query",
            "name": "registrationType",
            "schema": {
              "type": "string",
              "enum": [
                "individual",
                "group"
              ]
            },
            "description": "Filter by registration type"
          },
          {
            "in": "query",
            "name": "registrationNumber",
            "schema": {
              "type": "string"
            },
            "description": "Filter by registration number (partial match)"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter registrations from this date"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter registrations until this date"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in registration number, name, or email"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "All registrations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "registrations": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AdminRegistration"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "currentPage": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            },
                            "totalCount": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "hasPrevPage": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/v1/registrations": {
      "get": {
        "summary": "Get user's registrations",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registrations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Registration"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new registration",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "registrationType"
                ],
                "properties": {
                  "registrationType": {
                    "type": "string",
                    "enum": [
                      "individual",
                      "group"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration created successfully"
          },
          "400": {
            "description": "User already has an active registration"
          }
        }
      }
    },
...
    "/api/v1/registrations/{id}/submit": {
      "post": {
        "summary": "Submit completed registration",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Registration ID or User ID (both are supported for convenience)"
          }
        ],
        "responses": {
          "200": {
            "description": "Registration submitted successfully"
          },
          "400": {
            "description": "Registration incomplete or payment pending"
          }
        }
      }
    },
...
    "/api/v1/registrations/{id}/guardian-info": {
      "put": {
        "summary": "Update guardian information step",
        "tags": [
          "Registration Steps"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Registration ID or User ID (both are supported for convenience)"
          }
        ],
        "responses": {
          "200": {
            "description": "Guardian information updated successfully"
          }
        }
      }
    },
...
    "/api/v1/registrations/{id}/status": {
      "get": {
        "summary": "Get registration status",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Registration ID or User ID (both are supported for convenience)"
          }
        ],
        "responses": {
          "200": {
            "description": "Registration status retrieved successfully"
          }
        }
      }
    },
...
  }
}